requirements.txt
streamlit==1.38.0
pypdf==4.3.1
openai==1.46.0
requests==2.32.3
reportlab==4.2.2
python-docx==1.1.2
